server/ai/cloth-segmentation-mini/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
Pillow>=8.0.0
opencv-python>=4.4.0
gdown>=4.4.0
numpy>=1.19.0 